Extending the github login to handle 2-factor auth.
In the case that github returns 422, it is likely that the token was already
issued for the given github_auth note and URL. In this case, this new
code will automatically extend the note with the hostname. This will prevent
future TAs from pulling their hair out if they authorize this token on multiple machines.