Fix a bug with the EPFL ldap lookup
When students have multiple results returned, often times
the GASPAR is incorrect. It is usually GASPAR@FOO, where FOO is some unit.

To get rid of this, this commit adds a check for this case. When it
encounters it, it returns the first result, with only one GASPAR, with
the @FOO stripped away if it is there.

Based on the times we've seen this case, this should take care of it.